Tried using float for the ao in the vertex attributes and.... this happened

+++ FabledRealms/src/Engine/Graphics/Mesh/Vertex.h
@@ -7,5 +7,5 @@ struct Vertex
 
 	glm::vec3 Normal;
 	glm::vec3 Tangent;
-	int AO;
+	float AO;
 };
